resolve conflict in dev-master and eeprom
